File cleanup and struct renaming
